@@ -1,8 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
     <OutputType>Exe</OutputType>
+    <UserSecretsId>dcc6122f-e4f5-4771-84e3-473ff43797b9</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup>
@@ -11,9 +12,11 @@
 
   <ItemGroup>
     <PackageReference Include="Appccelerate.StateMachine" Version="4.4.0" />
+    <PackageReference Include="ioBroker.net" Version="0.0.15" />
     <PackageReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.1.2" PrivateAssets="All" />
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.9" />
+    <PackageReference Include="NSubstitute" Version="4.2.2" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc2" />
     <PackageReference Include="System.Device.Gpio" Version="0.1.0-prerelease.19310.5" />
     <PackageReference Include="UnitsNet" Version="4.26.0" />
