@@ -1,26 +1,33 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Device.Gpio;
 using System.Linq;
 using System.Threading.Tasks;
 using FlushTheToiletWebServer.CF01;
 using FlushTheToiletWebServer.Models;
 using FlushTheToiletWebServer.Services;
+using ioBroker.net;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.OpenApi.Models;
+using NSubstitute;
 
 namespace FlushTheToiletWebServer
 {
     public class Startup
     {
-        public Startup(IConfiguration configuration)
+        private readonly IWebHostEnvironment _env;
+
+        public Startup(IConfiguration configuration, IWebHostEnvironment env)
         {
             Configuration = configuration;
+            _env = env;
         }
 
         public IConfiguration Configuration { get; }
@@ -28,7 +35,8 @@ namespace FlushTheToiletWebServer
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            //services.AddMvc();
+            services.AddControllers();
 
             RegisterCf01Modules(services);
             RegisterServices(services);
@@ -43,7 +51,7 @@ namespace FlushTheToiletWebServer
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             // Enable middleware to serve generated Swagger as a JSON endpoint.
             app.UseSwagger();
@@ -60,7 +68,12 @@ namespace FlushTheToiletWebServer
                 app.UseDeveloperExceptionPage();
             }
 
-            app.UseMvc();
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
 
             // start state machine
             var stateMachine = app.ApplicationServices.GetService<IToiletFlusherStateMachineModel>();
@@ -69,6 +82,14 @@ namespace FlushTheToiletWebServer
 
         private void RegisterCf01Modules(IServiceCollection services)
         {
+            if (_env.IsDevelopment())
+            {
+                services.AddSingleton<IGpioController>(_ => Substitute.For<IGpioController>());
+            }
+            else
+            {
+                services.AddSingleton<IGpioController, GpioController>();
+            }
             services.AddSingleton<IFlusherMotor, FlusherMotor>();
             services.AddSingleton<ILedControl, LedControl>();
             services.AddSingleton<IManDetectionDistanceSensor, ManDetectionDistanceSensor>();
@@ -79,6 +100,7 @@ namespace FlushTheToiletWebServer
             services.AddSingleton<IFlushService, FlushService>();
             services.AddSingleton<IManDetectionService, ManDetectionService>();
             services.AddSingleton<IToiletFlusherStateMachine, ToiletFlusherStateMachine>();
+            services.AddSingleton<IIoBrokerDotNet>(_ => new IoBrokerDotNet(@"http://iobroker:8084"));
         }
 
         private void RegisterModels(IServiceCollection services)
