@@ -8,10 +8,10 @@ namespace FlushTheToiletWebServer.CF01.Devices
 {
     public class HCSR04
     {
-        private GpioController gpio = new GpioController();
+        private readonly IGpioController _gpio;
 
-        private int trig_Pin;
-        private int echo_Pin;
+        private int _trig_Pin;
+        private int _echo_Pin;
 
 
         static Object deviceLock = new object();
@@ -26,8 +26,9 @@ namespace FlushTheToiletWebServer.CF01.Devices
         /// <param name="trig_Pin"></param>
         /// <param name="echo_Pin"></param>
         /// <param name="timeoutMilliseconds">defaults to 20</param>
-        public HCSR04(byte trig_Pin, byte echo_Pin, int timeoutMilliseconds=20)
+        public HCSR04(IGpioController gpio, byte trig_Pin, byte echo_Pin, int timeoutMilliseconds=20)
         {
+            _gpio = gpio;
             Initialise(trig_Pin, echo_Pin, timeoutMilliseconds);
         }
 
@@ -37,23 +38,24 @@ namespace FlushTheToiletWebServer.CF01.Devices
         /// <param name="trig_Pin"></param>
         /// <param name="echo_Pin"></param>
         /// <param name="maxDistance">Set Ultra Sonic maximum distance to detect.  This is approximate only.  Based on 34.3 cm per millisecond, 20 degrees C at sea level</param>
-        public HCSR04(byte trig_Pin, byte echo_Pin, Length maxDistance)
+        public HCSR04(IGpioController gpio, byte trig_Pin, byte echo_Pin, Length maxDistance)
         {
+            _gpio = gpio;
             int milliSeconds = (int)(maxDistance.Centimeters / 34.3 * 2);
             Initialise(trig_Pin, echo_Pin, milliSeconds);
         }
 
         private void Initialise(byte trig_Pin, byte echo_Pin, int timeoutMilliseconds)
         {
-            this.trig_Pin = trig_Pin;
-            this.echo_Pin = echo_Pin;
+            _trig_Pin = trig_Pin;
+            _echo_Pin = echo_Pin;
 
             TimeoutMilliseconds = timeoutMilliseconds;
 
-            gpio.OpenPin(trig_Pin, PinMode.Output);
-            gpio.OpenPin(echo_Pin, PinMode.Input);
+            _gpio.OpenPin(trig_Pin, PinMode.Output);
+            _gpio.OpenPin(echo_Pin, PinMode.Input);
             
-            gpio.Write(trig_Pin, PinValue.Low);
+            _gpio.Write(trig_Pin, PinValue.Low);
         }
 
         /// <summary>
@@ -86,14 +88,14 @@ namespace FlushTheToiletWebServer.CF01.Devices
                 //http://www.modmypi.com/blog/hc-sr04-ultrasonic-range-sensor-on-the-raspberry-pi
 
 
-                gpio.Write(trig_Pin, PinValue.Low);                   // ensure the trigger is off
+                _gpio.Write(_trig_Pin, PinValue.Low);                   // ensure the trigger is off
                 Task.Delay(TimeSpan.FromMilliseconds(1)).Wait();  // wait for the sensor to settle
 
-                gpio.Write(trig_Pin, PinValue.High);                       // turn on the pulse
+                _gpio.Write(_trig_Pin, PinValue.High);                       // turn on the pulse
                 Task.Delay(TimeSpan.FromMilliseconds(.01)).Wait();      // let the pulse run for 10 microseconds
-                gpio.Write(trig_Pin, PinValue.Low);                        // turn off the pulse
+                _gpio.Write(_trig_Pin, PinValue.Low);                        // turn off the pulse
 
-                var time = PulseIn(echo_Pin, PinValue.High, timeoutMilliseconds);
+                var time = PulseIn(_echo_Pin, PinValue.High, timeoutMilliseconds);
 
                 // https://en.wikipedia.org/wiki/Speed_of_sound
                 // speed of sound is 34300 cm per second or 34.3 cm per millisecond
@@ -109,7 +111,7 @@ namespace FlushTheToiletWebServer.CF01.Devices
             sw.Restart();
 
             // Wait for pulse
-            while (sw.ElapsedMilliseconds < timeout && gpio.Read(pin_number) != value) { }
+            while (sw.ElapsedMilliseconds < timeout && _gpio.Read(pin_number) != value) { }
 
             if (sw.ElapsedMilliseconds >= timeout)
             {
@@ -119,7 +121,7 @@ namespace FlushTheToiletWebServer.CF01.Devices
             sw.Restart();
 
             // Wait for pulse end
-            while (sw.ElapsedMilliseconds < timeout && gpio.Read(pin_number) == value) { }
+            while (sw.ElapsedMilliseconds < timeout && _gpio.Read(pin_number) == value) { }
 
             sw.Stop();
 
