@@ -1,6 +1,7 @@
 ﻿using FlushTheToiletWebServer.CF01.Devices;
 using System;
 using System.Collections.Generic;
+using System.Device.Gpio;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -15,9 +16,9 @@ namespace FlushTheToiletWebServer.CF01
 
         private HCSR04 mDistanceSensor;
 
-        public ManDetectionDistanceSensor()
+        public ManDetectionDistanceSensor(IGpioController gpio)
         {
-            mDistanceSensor = new HCSR04(TRIGGER_PIN, ECHO_PIN, Length.FromMeters(3.0));
+            mDistanceSensor = new HCSR04(gpio, TRIGGER_PIN, ECHO_PIN, Length.FromMeters(3.0));
         }
 
         public Length GetDistance()
